moved transactionable functionality to a concern
