+++ app/models/purchase.rb
@@ -16,12 +16,11 @@
 class Purchase < ActiveRecord::Base
   acts_as_paranoid
 
+  include Transactionable
+
   store :meta_data, accessors: [], coder: Hash
 
-  has_many :transaction_items, as: :transactionable
   belongs_to :creator, class_name: Person, foreign_key: 'person_id'
-
-  accepts_nested_attributes_for :transaction_items, allow_destroy: true
 
   validates :purchase_date, :person_id, :vendor, :transaction_items, presence: true
 
@@ -33,21 +32,4 @@ class Purchase < ActiveRecord::Base
 
   after_create :add_to_stock
   before_destroy :remove_from_stock
-
-  private
-    def add_to_stock
-      self.transaction_items.each do |transaction_item|
-        item = transaction_item.item
-        item.update_attribute(:stock_quantity,
-          item.stock_quantity + transaction_item.quantity)
-      end
-    end
-
-    def remove_from_stock
-      self.transaction_items.each do |transaction_item|
-        item = transaction_item.item
-        item.update_attribute(:stock_quantity,
-          item.stock_quantity - transaction_item.quantity)
-      end
-    end
 end
